missed int -> NSInteger transition
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gdl2/trunk@36243 72102866-910b-0410-8b05-ffd578937521

+++ GDL2Palette/DisplayGroupInspector.m
@@ -104,7 +104,10 @@ objectValueForTableColumn:(NSTableColumn *)tc
   return [_localKeys objectAtIndex:row];
 }
 
-- (void) tableView:(NSTableView *)tv setObjectValue:(id)newValue forTableColumn:(NSTableColumn *)tc row:(int) row;
+- (void) tableView:(NSTableView *)tv 
+setObjectValue:(id)newValue
+forTableColumn:(NSTableColumn *)tc
+	   row:(NSInteger) row;
 {
   [_localKeys replaceObjectAtIndex:row withObject:newValue];
   [object setLocalKeys:_localKeys];
